Move server & server implementations to own package; leave interfaces and exceptions in shared 'interfaces' package
