+++ src/main/java/net/teamtrycatch/server/AccountDatastoreImpl.java
@@ -1,6 +1,9 @@
 package net.teamtrycatch.server;
 
-import net.teamtrycatch.shared.AccountNotFoundException;
+import net.teamtrycatch.shared.interfaces.AccountNotFoundException;
+import net.teamtrycatch.shared.server.DepositTransaction;
+import net.teamtrycatch.shared.server.InitialTransaction;
+import net.teamtrycatch.shared.server.WithdrawalTransaction;
 
 import java.text.DateFormat;
 import java.text.ParseException;
