+++ src/main/java/net/teamtrycatch/server/PersonalAccount.java
@@ -1,6 +1,9 @@
 package net.teamtrycatch.server;
 
-import net.teamtrycatch.shared.Transaction;
+import net.teamtrycatch.shared.server.DepositTransaction;
+import net.teamtrycatch.shared.server.InitialTransaction;
+import net.teamtrycatch.shared.server.WithdrawalTransaction;
+import net.teamtrycatch.shared.interfaces.Transaction;
 
 import javax.crypto.SecretKeyFactory;
 import javax.crypto.spec.PBEKeySpec;
