--- src/main/java/net/teamtrycatch/shared/IllegalArguementException.java
@@ -1,11 +1,0 @@
-package net.teamtrycatch.shared;
-
-public class IllegalArguementException extends Exception {
-    public IllegalArguementException(String message) {
-        super(message);
-    }
-
-    public IllegalArguementException() {
-    	
-    }
-}
