+++ src/main/java/net/teamtrycatch/shared/client/IllegalArgumentException.java
@@ -1,0 +1,11 @@
+package net.teamtrycatch.shared.client;
+
+public class IllegalArgumentException extends Exception {
+    public IllegalArgumentException(String message) {
+        super(message);
+    }
+
+    public IllegalArgumentException() {
+    	
+    }
+}
