+++ src/main/java/net/teamtrycatch/shared/interfaces/InvalidSession.java
@@ -1,4 +1,4 @@
-package net.teamtrycatch.shared;
+package net.teamtrycatch.shared.interfaces;
 
 public class InvalidSession extends Exception {
     public InvalidSession(String message) {
