+++ src/main/java/net/teamtrycatch/shared/server/WithdrawalTransaction.java
@@ -1,6 +1,6 @@
-package net.teamtrycatch.server;
+package net.teamtrycatch.shared.server;
 
-import net.teamtrycatch.shared.Transaction;
+import net.teamtrycatch.shared.interfaces.Transaction;
 
 import java.util.Date;
