+++ src/test/java/net/teamtrycatch/server/PersonalAccountTest.java
@@ -1,6 +1,9 @@
 package net.teamtrycatch.server;
 
-import net.teamtrycatch.shared.Transaction;
+import net.teamtrycatch.shared.server.DepositTransaction;
+import net.teamtrycatch.shared.server.InitialTransaction;
+import net.teamtrycatch.shared.server.WithdrawalTransaction;
+import net.teamtrycatch.shared.interfaces.Transaction;
 import org.junit.Before;
 import org.junit.Test;
